requirements.txt
langchain
openai
chromadb
python-dotenv
pandas